% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainImageClass.R
\name{checkClass}
\alias{checkClass}
\title{Checks if object is of given class and stops 
if not.}
\usage{
checkClass(obj, expClass)
}
\arguments{
\item{obj}{Object to test}

\item{expClass}{character of the class}
}
\value{
TRUE if everything is ok. Otherwise, execution is halted.
}
\description{
An object obj is tests if its class 
equals expClass. If not, execution is halted.
}
\examples{
data <- data.frame(x=1:10, y=1:10, 
z=c(rep(2,5), rep(3,5)), val=rnorm(100))
img <- new("brainImage", data, "test", 10)
checkClass(img, "brainImage")
}

