% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dec}
\alias{dec}
\title{Decode coordinates}
\usage{
dec(val)
}
\arguments{
\item{val}{encoded coordinates}
}
\value{
vector containing the (x,y,z) coordinates
}
\description{
Extracts (x,y,z) values fro msingle value
}
\examples{
dec(1050627)
}

