% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{dec2enc}
\alias{dec2enc}
\title{Convert DEC to ENC}
\usage{
dec2enc(data, index = 4)
}
\arguments{
\item{data}{data.frame for volume (x,y,z, val)}

\item{index}{colum with values}
}
\value{
data.frame containing values and encoded coordinates
}
\description{
Convert (x,y,z) coordinates to encoded coordinated
}
\examples{
data <- data.frame(
x=c(60,61,62),
y=c(40,40,40),
z=c(1,1,1),
val=c(234,112,45)
)
options("BITSIZE"=10)
dec2enc(data)
}

