% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeCOORD.R
\name{decodeCOORD}
\alias{decodeCOORD}
\title{Convert ENC to DEC}
\usage{
decodeCOORD(enc)
}
\arguments{
\item{enc}{value, containing x,y,z, values}

\item{values}{imaging values}
}
\value{
data.frame with the decoded coordiantes
}
\description{
Converts encoded coordinates to x,y,z values
}
\examples{
enc<-c(51512331,51512254,51513355)
val<-c(342,221,123)
options("BITSIZE"=10)
enc2dec(val,enc)
}

