% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeDecodeXYZCOORD.R
\name{decodeCOORDtoXYZ}
\alias{decodeCOORDtoXYZ}
\title{Decode coordinates}
\usage{
decodeCOORDtoXYZ(val)
}
\arguments{
\item{val}{encoded coordinates}
}
\value{
vector containing the (x,y,z) coordinates
}
\description{
Extracts (x,y,z) values fro msingle value
}
\examples{
decodeCOORDtoXYZ(1050627)
}

