% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{decX}
\alias{decX}
\title{Extracts x-values}
\usage{
decX(val)
}
\arguments{
\item{val}{encoded coordinate value}
}
\value{
decoded x coordinate
}
\description{
Extracts x values from encoded number
}
\examples{
decX(1050627)
}

