% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{decZ}
\alias{decZ}
\title{Extracts z-values}
\usage{
decZ(val)
}
\arguments{
\item{val}{encoded coordinate value}
}
\value{
decoded z coordinate
}
\description{
Extracts z values from encoded number
}
\examples{
decZ(1050627)
}

