% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrFast2.R, R/getMinimaKeyVal.R
\name{density.adapt}
\alias{density.adapt}
\title{Helper function for function approximation}
\usage{
\method{density}{adapt}(x, N, minV, maxV)

\method{density}{adapt}(x, N, minV, maxV)
}
\arguments{
\item{x}{data to approximate fun for}

\item{minV}{minimum value of x}

\item{maxV}{maximum value of x}

\item{length}{of x}

\item{x}{data to approximate fun for}

\item{length}{of x}

\item{minV}{minimum value of x}

\item{maxV}{maximum value of x}
}
\value{
Function approximating the initial data

Function approximating the initial data
}
\description{
Combines density and approxfun. 
WARNING: All data consistency checks have been removed!

Combines density and approxfun. 
WARNING: All data consistency checks have been removed!
}
\examples{
data <- rnorm(100)
minV <- min(data)
maxV <- max(data)
n <- length(data)
density.adapt(data, n, minV, maxV)
data <- rnorm(100)
minV <- min(data)
maxV <- max(data)
n <- length(data)
density.adapt(data, n, minV, maxV)
}

