% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{enc}
\alias{enc}
\title{Encode coordinates}
\usage{
enc(x, y, z)
}
\arguments{
\item{x}{x-values}

\item{y}{y-values}

\item{z}{z-values}
}
\value{
encoded value of an (x,y,z) coord tupel
}
\description{
Enoces (x,y,z) coordinates
}
\examples{
enc(1,2,3)
}

