% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{enc2dec}
\alias{enc2dec}
\title{Convert ENC to DEC}
\usage{
enc2dec(values, enc)
}
\arguments{
\item{values}{imaging values}

\item{enc}{value, containing x,y,z, values}
}
\value{
data.frame with the decoded coordiantes
}
\description{
Converts encoded coordinates to x,y,z values
}
\examples{
enc<-c(51512331,51512254,51513355)
val<-c(342,221,123)
options("BITSIZE"=10)
enc2dec(val,enc)
}

