% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr.R
\name{extr}
\alias{extr}
\title{Find extremal values}
\usage{
extr(data, zeichne = TRUE, delta = 0.01)
}
\arguments{
\item{data}{vector of values to analyze}

\item{zeichne}{draw density and identified values}

\item{delta}{steps between analyzed points for the fitted density curve}
}
\value{
list, containing two data.frames: min and max key/value pairs
}
\description{
Finds minima and maxima.
}
\examples{
extr(c(1:10,2,3,4,1:10,10:1,2:4,1:10), zeichne = FALSE, delta = 0.5)
}

