% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrFast2.R
\name{extrFast2}
\alias{extrFast2}
\title{Retrieve minimal values}
\usage{
extrFast2(data, minV, maxV, delta)
}
\arguments{
\item{data}{vector af values to analyze}

\item{minV}{minimum value of data vector (performance)}

\item{maxV}{maximim value of data vector (performance)}

\item{delta}{steps to analyze the resulting fitted fun}
}
\value{
list, containing a data.frame with the columns 
key and val (minima).
}
\description{
Retrieve minimal values, parameters are set outside (!)
}
\examples{
data <- rnorm(100)
minV <- min(data)
maxV <- max(data)
extrFast2(data, minV, maxV, 1)
}

