% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxDensityFun.R
\name{fitDensityFunGetX}
\alias{fitDensityFunGetX}
\title{Helper function for density function approximation}
\usage{
fitDensityFunGetX(x, N, minV, maxV)
}
\arguments{
\item{x}{data to approximate fun for}

\item{N}{length of x}

\item{minV}{minimum value of x}

\item{maxV}{maximum value of x}
}
\value{
Function approximating the initial data
}
\description{
Combines density and approxfun. 
WARNING: All data consistency checks have been removed!
}
\examples{
data <- rnorm(100)
minV <- min(data)
maxV <- max(data)
n <- length(data)
fitDensityFunGetX(data, n, minV, maxV)
}

