% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainImageClass.R
\name{getCoords}
\alias{getCoords}
\title{Get all encoded coordinates}
\usage{
getCoords(img)
}
\arguments{
\item{img}{brainImage instance}
}
\value{
COORD vector
}
\description{
Encoded coordinates (COORD vector)
}
\examples{
data <- data.frame(x=1:10, y=1:10, 
z=c(rep(2,5), rep(3,5)), val=rnorm(100))
img <- new("brainImage", data, "test", 10)
getCoords(img)
}

