% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImage.R
\name{getImage}
\alias{getImage}
\title{Returns an image instance}
\usage{
getImage(data)
}
\arguments{
\item{data}{data.frame with the columns (x,y,z,values)}
}
\value{
image object
}
\description{
Creates an Image object from a (x,y,z,val) data.frame
}
\examples{
data <- data.frame(x=1:10, y=1:10, 
z=c(rep(2,5), rep(3,5)), val=rnorm(100))
getImage(data)
}

