% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainImageClass.R
\name{getMeasurements}
\alias{getMeasurements}
\title{Get all values as vector}
\usage{
getMeasurements(img)
}
\arguments{
\item{img}{brainImage instance}
}
\value{
measurements vector
}
\description{
All measurements stored in the img object
are accessed
}
\examples{
data <- data.frame(x=1:10, y=1:10, 
z=c(rep(2,5), rep(3,5)), val=rnorm(100))
img <- new("brainImage", data, "test", 10)
getMeasurements(img)
}

