% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrFast2.R
\name{getMinimaKeyVal}
\alias{getMinimaKeyVal}
\title{Retrieve all minima from a given data vector}
\usage{
getMinimaKeyVal(data, n, minV, maxV, delta)
}
\arguments{
\item{data}{vector af values to analyze}

\item{n}{length of data}

\item{minV}{minimum value of data vector (performance)}

\item{maxV}{maximim value of data vector (performance)}

\item{delta}{steps to analyze the resulting fitted fun}
}
\value{
list, containing a data.frame with the columns 
key and val (minima).
}
\description{
Retrieve minimal values
}
\examples{
data <- rnorm(100)
minV <- min(data)
maxV <- max(data)
getMinimaKeyVal(data, minV, maxV, 1)
}

