% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainImageClass.R
\name{getNumberOfMeasurements}
\alias{getNumberOfMeasurements}
\title{Get Number of measurements (non NA)}
\usage{
getNumberOfMeasurements(img)
}
\arguments{
\item{img}{brainImage instance}
}
\value{
number of non-NA measurements
}
\description{
Returns the total amount of measurements
which are not NA in this image
}
\examples{
data <- data.frame(x=1:10, y=1:10, 
z=c(rep(2,5), rep(3,5)), val=rnorm(100))
img <- new("brainImage", data, "test", 10)
getNumberOfMeasurements(img)
}

