% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubclasses.R
\name{getSubclasses}
\alias{getSubclasses}
\title{Extract Subclass data}
\usage{
getSubclasses(data, split, index = 4)
}
\arguments{
\item{data}{data.frame of volume to split}

\item{split}{data.frame (key, val) with values to split at (keys)}

\item{index}{column index with values in the volume dataset}
}
\value{
returns a list where all list elements contain 
data subsets belonging to the identified tissue class
}
\description{
Splits a given dataset into subclasses.
}
\examples{
data <- data.frame(x=rep(1:10, 5), 
y=c(rep(1,10),rep(2,10),rep(3,10),rep(4,10),rep(5, 10)), 
z=rep(1, 50), val=rnorm(50))
getSubclasses(data, data.frame(key=median(data$val), val=NA))
}

