% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainImageClass.R
\name{getThresholds}
\alias{getThresholds}
\title{Get tissue class thresholds}
\usage{
getThresholds(img)
}
\arguments{
\item{img}{brainImage instance}
}
\value{
adjusted imageBrain object
}
\description{
Returns cutoffs for different tissue 
classes, identified with 
identifyMinima(getMeasurements(brainImage), iterat=10000)
and set with setThresholds().
}
\examples{
options("BITSIZE"=10)
data <- data.frame(x=1:15, y=1:15, 
z=c(rep(2,15), rep(3,15)), val=rnorm(6750))
img <- new("brainImage", data, "test", 10)
img <- setThresholds(img,
imageanalysisBrain::identifyMinima(getMeasurements(img), iterat=1000))
getThresholds(img)
}

