% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTissueSubclass.R
\name{getTissueSubclasses}
\alias{getTissueSubclasses}
\title{Create images for each identified tissue subclass.}
\usage{
getTissueSubclasses(data, split)
}
\arguments{
\item{data}{data.frame of volume to split}

\item{split}{data.frame (key, val) with values to split at (keys)}
}
\value{
returns a list where all list elements contain 
data subsets belonging to the identified tissue class
}
\description{
Creates brainImage instances for each subclass 
preserving the original dimensions.
}
\examples{
data <- data.frame(x=rep(1:10, 5), 
y=c(rep(1,10),rep(2,10),rep(3,10),rep(4,10),rep(5, 10)), 
z=rep(1, 50), val=rnorm(50))
img <- new("brainImage", data, "test", 10)
getTissueSubclasses(img, data.frame(key=median(data$val), val=NA))
}

