% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainImageClass.R
\name{getXFactor}
\alias{getXFactor}
\title{Get X-Scaling factor}
\usage{
getXFactor(img)
}
\arguments{
\item{img}{brainImage instance}
}
\value{
scaling factor
}
\description{
Factor to normalize the maximum biparietal 
dimension to 1. Default: NULL. Initialize with 
calcXScalingFactor
}
\examples{
data <- data.frame(x=1:10, y=1:10, 
z=c(rep(2,5), rep(3,5)), val=rnorm(100))
img <- new("brainImage", data, "test", 10)
img <- imageanalysisBrain::calcXScalingFactor(img)
getXFactor(img)
}

