% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeDecodeXYZCOORD.R
\name{getXFromCOORD}
\alias{getXFromCOORD}
\title{Extracts x-values}
\usage{
getXFromCOORD(val)
}
\arguments{
\item{val}{encoded coordinate value}
}
\value{
decoded x coordinate
}
\description{
Extracts x values from encoded number
}
\examples{
decX(1050627)
}

