% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXRange.R
\name{getXRange}
\alias{getXRange}
\title{Find largest X-Dimension value}
\usage{
getXRange(data, steps = 500)
}
\arguments{
\item{data}{data.frame of volume for which the maximal 
x dimension should be retrieved}

\item{steps}{use every n. value (n=steps)}
}
\value{
list containing the minimal x and maximal x values
}
\description{
Needed for scaling due to different resolutions.
The Scaling value is calculated to set the maximum 
X-dimension to 1
}
\examples{
data <- data.frame(x=rep(1:10, 5), 
y=c(rep(1,10),rep(2,10),rep(3,10),rep(4,10),rep(5, 10)), 
z=rep(1, 50), rnorm(50))
getXRange(data)
}

