% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZeroKoord.R
\name{getZeroKoord}
\alias{getZeroKoord}
\title{Find points of origins}
\usage{
getZeroKoord(data, sparse = 500, draw = FALSE, col = "yellow",
    drawsteps = 15)
}
\arguments{
\item{data}{data frame of volume for which the center or origin 
should be calculated}

\item{sparse}{select only each n. voxel (n=sparse) to speed up calculations}

\item{draw}{draw the identified planes}

\item{col}{color of planes}

\item{drawsteps}{select only each n. voxel (n=drawsteps) to draw as 
identified plane}
}
\value{
coordinates belonging to the point of origin (Zero Coordinate)
}
\description{
Determines the point of origin for a given volume by
first selecting the
}
\examples{
data <- data.frame(x=rep(1:10, 5), 
y=c(rep(1,10),rep(2,10),rep(3,10),rep(4,10),rep(5, 10)), 
z=rep(1, 50), rnorm(50))
data$COORD <- encodeXYZToCOORD(data$x, data$y, data$z)
getZeroKoord(data, 1)
}

