% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\name{identifyMinima}
\alias{identifyMinima}
\title{identifies minima in a given set of values}
\usage{
identifyMinima(data, proportion = 0.001, iterat = 5e+05, delta = 1)
}
\arguments{
\item{data}{vector of values to analyze}

\item{proportion}{proportion of values to analyze 
(proportions*length(data)), if below 10, then the value 
is mutiplied by 10 in each iteration}

\item{iterat}{number of sample drawings}

\item{delta}{width between two points to differentiate}
}
\value{
data.frame containing all identified minima
}
\description{
Identifies minima by repeatedly drawing 
small numbers of values from a larger set.
}
\examples{
identifyMinima(1:1000, 1, 10)
}

