\name{imageanalysisBrain-package}
\alias{imageanalysisBrain-package}
\alias{imageanalysisBrain}
\docType{package}
\title{
\packageTitle{imageanalysisBrain}
}
\description{
\packageDescription{imageanalysisBrain}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{imageanalysisBrain}
\packageIndices{imageanalysisBrain}
}
\author{
\packageAuthor{imageanalysisBrain}

Maintainer: \packageMaintainer{imageanalysisBrain}
}
\keyword{ package }
