% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainImageClass.R
\docType{methods}
\name{initialize,brainImage-method}
\alias{initialize,brainImage-method}
\title{Constructor brainImage}
\usage{
\S4method{initialize}{brainImage}(.Object, values = data.frame,
  filename = character, coordBIT = numeric, ...)
}
\value{
a new brainImage instance
}
\description{
Constructor for new brainImage instances
}

