% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZSlice.R
\name{plotZSlice}
\alias{plotZSlice}
\title{Plot z slice of brainImage instance}
\usage{
plotZSlice(data, z, index = 4, ret = FALSE, add = FALSE,
  col = grDevices::gray.colors(1024), main = "", dimV = NULL)
}
\arguments{
\item{data}{data.frame with x,y,z,val columns}

\item{z}{selected z slice to plot}

\item{index}{column with values to plot}

\item{ret}{return the image (T,F) as matrix}

\item{add}{add to existing plot}

\item{col}{color palette}

\item{main}{main}

\item{minx}{min X-value plotted}

\item{maxx}{max X-value plotted}

\item{miny}{min Y value plotted}

\item{maxy}{max Y value plotted}
}
\value{
list of min/max values used for this plot
}
\description{
Plot given z slice of brainImage instance
}
\examples{
data <- data.frame(x=rep(1:10, 5), 
y=c(rep(1,10),rep(2,10),rep(3,10),rep(4,10),rep(5, 10)), 
z=rep(1, 50), rnorm(50))
plotZSlice(data, 1)
}

