% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\name{sampleData}
\alias{sampleData}
\title{Sample data}
\usage{
sampleData(data, proportion = 1.001, iterat = 5e+05, delta = 100)
}
\arguments{
\item{data}{vector of values to analyze}

\item{proportion}{proportion of voxels to analyze 
(proportions*length(data)), if below 10, then the value 
is mutiplied by 10}

\item{iterat}{number of sample drawings}
}
\value{
data.frame containing all identified minima
}
\description{
Fast alternative for extr(), returning only minima
and using parallelization
}
\examples{
sampleData(1:1000, 1, 10)
}

