% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCoordToNA.R
\name{setCoordToNA}
\alias{setCoordToNA}
\title{Set selected coordinates to NA}
\usage{
setCoordToNA(data, sel)
}
\arguments{
\item{data}{brainImage instance to manipulate}

\item{sel}{vector of positions (sequential) which
should be set to NA}
}
\value{
brainImage instance with adapted values
}
\description{
When creating subclass images, values not belonging
to the respective class can be set to NA with this function
}

