% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtremalValues.R
\name{getExtremalValues}
\alias{getExtremalValues}
\title{Find extremal values (min+max).}
\usage{
getExtremalValues(data, zeichne = TRUE, delta = 0.01)
}
\arguments{
\item{data}{vector of values to analyze}

\item{zeichne}{draw density and identified values}

\item{delta}{steps between analyzed points for the fitted density curve}
}
\value{
list, containing two data.frames: min and max key/value pairs
}
\description{
Finds minima and maxima.
}
\examples{
getExtremalValues(c(1:10,2,3,4,1:10,10:1,2:4,1:10), zeichne = FALSE, delta = 0.5)
}

