% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getYFromCOORD}
\alias{getYFromCOORD}
\title{Extracts y-values}
\usage{
getYFromCOORD(val)
}
\arguments{
\item{val}{encoded coordinate value}
}
\value{
decoded y coordinate
}
\description{
Extracts y values from encoded number
}
\examples{
decY(1050627)
}

