% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainImageClass.R
\name{setThresholds}
\alias{setThresholds}
\title{Set tissue class thresholds}
\usage{
setThresholds(img, thresholds)
}
\arguments{
\item{img}{brainImage instance}

\item{thresholds}{data.frame with key/val columns,
calculated with identifyMinima()}
}
\value{
adjusted imageBrain object
}
\description{
Sets tissue class thresholds for this 
object. Thresholds can be caluclated with 
identifyMinima().
}
\examples{
data <- data.frame(x=1:15, y=1:15, 
z=c(rep(2,15), rep(3,15)), val=rnorm(6750))
img <- new("brainImage", data, "test", 10)
img <- setThresholds(img,
imageanalysisBrain::identifyMinima(getMeasurements(img), iterat=1000))
}

